Updates performed: - Code refactoring; - Changed routes; - Optimized exception message generation when object doesn't exist; - Added ObjectNotExist exceptions tests.

+++ ReadingList.Tests/ReadingTimeCalculationTests.cs
@@ -25,7 +25,7 @@ namespace ReadingList.Tests
                               ReadingTimeCalculator.Calculate(item.Status, item.LastStatusUpdateDate,
                                   BookItemStatus.Reading);
 
-            Assert.True(readingTime == default(double));
+            Assert.True(readingTime == default(int));
         }
 
         [Fact]
